Added an alternate destination path option
Added an option to provide a backup destination path for your backups.
This allows you to specify a mounted drive or volume for your backup
process. However, if you backup volume isn't available for some reason
the backup will revert to an alternate path where the files can be
manually dealt with at a later date.