Improving the date checking
This change does a much better job at looking for old snapshots to
delete. It compares the unix timestamps of each file from the filename
date string and then deletes the old ones. Relies upon the C script in
this package. This is needed to make the date comparison work on BSD
systems where the date -d argument doesn't exist.